+++ woocommerce.php
@@ -10,6 +10,51 @@
 
 get_header(); 
 
+if(is_shop() && false) {	?>
+<div class="main-container full-width">
+	<div class="main-grid">
+		<div class="main-content-full-width">
+		<?php
+		get_template_part('template-parts/shop-page-featuredcat');
+		get_template_part('template-parts/shop-page-bookcats');
+		get_template_part('template-parts/shop-page-merchcat');
+		?>
+		</div>
+	</div>
+</div>
+<?php
+	// Function to trigger the slick slider on the shop page.
+	if ( !function_exists('bp_shop_page_slider') ) {
+
+		add_action( 'wp_footer', 'bp_shop_page_slider', 100 );
+		function bp_shop_page_slider() {
+			?><script type="text/javascript"> 
+				
+				$("#showcase-slider").slick( {
+					arrows: false,
+					slidesToShow: 6,
+					adaptiveHeight: true,
+					responsive: [{
+						breakpoint: 900,
+						settings: {
+							slidesToShow: 3,
+							slidesToScroll: 3,
+							infinite: true,
+						}},
+						{
+						breakpoint: 600,
+						settings: {
+							slidesToShow: 3,
+							slidesToScroll: 3,
+							infinite: true,
+						}
+					}]
+				});			
+			</script><?php
+		}	
+	}
+}
+else {
 ?> 
 <div class="main-container">
 	<div class="main-grid">
@@ -24,6 +69,5 @@ get_header();
 	<?php get_sidebar(); ?>
 	</div>
 </div>
-<?php 
-
+<?php }
 get_footer();